Time: 181 ms (15.26%), Space: 130.9 MB (40%) - LeetHub

+++ 0325-maximum-size-subarray-sum-equals-k/0325-maximum-size-subarray-sum-equals-k.cpp
@@ -2,13 +2,13 @@ class Solution {
 public:
     int maxSubArrayLen(vector<int>& nums, int k) {
         int n = nums.size();
-        unordered_map<long,long> dict;
+        unordered_map<long,int> dict;
         long sum=0,res=0;
         dict[0]=-1;
         for(int i=0;i<n;i++){
             sum+=nums[i];
             if(dict.find(sum-k)!=dict.end()){
-                res=max(res,i-dict[sum-k]);
+                res=max(res,(long)(i-dict[sum-k]));
             }
             if(dict.find(sum)==dict.end()){
                 dict[sum]=i;
